Create a new folder for each component with style
