+++ src/components/CurryImagePicker/style.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from "react-native";
+export const styles = StyleSheet.create({
+    container:{
+        width: '100%',
+        alignItems: 'center',
+    },
+    imageContainer:{
+        borderWidth:1,
+        borderColor: 'black',
+        width:'80%',
+        height: 180,
+        backgroundColor:'#eee',
+    },
+    button:{
+        margin:8
+    }
+   
+});
