+++ src/components/DatePicker/style.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native";
+export const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+  });
