+++ src/components/ImageViewer/style.js
@@ -1,0 +1,19 @@
+import React from "react";
+export const styles = StyleSheet.create({
+    container:{
+        width: '100%',
+        alignItems: 'center',
+        
+    },
+    imageContainer:{
+        borderWidth:1,
+        borderColor: 'black',
+        width:'80%',
+        height: 180,
+        backgroundColor:'#eee',
+    },
+    button:{
+        margin:8
+    }
+   
+})
